trasladar datos comunes de personas a clase Persona y hacer Estudiante heredar de persona.
