+++ src/main/java/org/example/model/Estudiante.java
@@ -2,18 +2,13 @@ package org.example.model;
 
 import java.util.Date;
 
-public class Estudiante {
+public class Estudiante extends Persona{
     private Long id;
-    private String nombre;
-    private String apellido;
-    private Date fechaDeNacimiento;
     private EstadoEstudiante estado;
 
     public Estudiante(Long id, String nombre, String apellido, Date fechaDeNacimiento, EstadoEstudiante estado) {
+        super(nombre, apellido, fechaDeNacimiento);
         this.id = id;
-        this.nombre = nombre;
-        this.apellido = apellido;
-        this.fechaDeNacimiento = fechaDeNacimiento;
         this.estado = estado;
     }
 
@@ -25,29 +20,6 @@ public class Estudiante {
         this.id = id;
     }
 
-    public String getNombre() {
-        return nombre;
-    }
-
-    public void setNombre(String nombre) {
-        this.nombre = nombre;
-    }
-
-    public String getApellido() {
-        return apellido;
-    }
-
-    public void setApellido(String apellido) {
-        this.apellido = apellido;
-    }
-
-    public Date getFechaDeNacimiento() {
-        return fechaDeNacimiento;
-    }
-
-    public void setFechaDeNacimiento(Date fechaDeNacimiento) {
-        this.fechaDeNacimiento = fechaDeNacimiento;
-    }
 
     public EstadoEstudiante getEstado() {
         return estado;
